+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Tapping on a tab bar icon can let you scroll to top.
+- Fixed a bug where a root note could be rendered as a reply
+- Added the option to copy the text content while browsing a note.
+- Fixed UI bugs when displaying the root note of replies.
 - Keep track of read stories.
 
 ## [0.1 (94)] - 2023-11-17Z