+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Keep track of read stories.
+- Fix an issue where reposts were not displaying correctly.
 
 ## [0.1 (94)] - 2023-11-17Z