+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added the option to copy the NIP-05 identifier when browsing a profile.
 - Tapping on a tab bar icon can let you scroll to top.
+- Fix an issue where reposts were not displaying correctly.
 
 ## [0.1 (96)] - 2023-11-28Z
 
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the option to copy the text content while browsing a note.
 - Fixed UI bugs when displaying the root note of replies.
 - Keep track of read stories.
+- Fix an issue where reposts were not displaying correctly.
 
 ## [0.1 (94)] - 2023-11-17Z